Change to use byteClasses for . character
The DFA state transition works by mapping the current character to a
byteClass. We handle this with a 128 element array, which we can access,
so long as the current character is an ASCII character. We can extend
this to support the . metacharacter, by mapping all characters > 127 to
a catchall character.