Restructure DFA transitions
First, we add additional state explicit state transitions back to
the zero state, removing the need to have our main loop check
for a -1 state. I.E. when matching 'abc' against 'adabc', we'll
go to state 0 upon seeing 'd', instead of -1.

Second, we change our array of state transitions from a 2d array to a
single array. Instead of looking up a sub-array using the current state,
we multiply the state by the number of byte classes to compute the index
we use.