Update indexForwards to use suffix searching.
This initial implementation is limited in three main ways, but still
provides a substantial speedup on many regexes.

First, we only use suffix matches when our regex has a maximum length,
so a regex like [a-z]+\.com will not use the .com suffix.

Second, we only execute suffix matches when we do not have a useful
prefix to match. Moreover, we don't combine the suffix match with
searching forwards using predicate tests.

Third, after executing the suffix match, we do not use that information
to decrease the number of states in our regex, instead we rematch the
suffix upon entering the appropriate states. Nor do we do anything like
search backwards from the suffix (this is related to the first
limitation).