Update outer loop to factor in min length of regex
Given a regex that has a min length of n, the outer loop will exit as
soon as we no longer have n characters to consume in our regex. This
doens't handle the case where we return to the zero state from a regex
that doesn't have a prefix or suffix.